+++ lib/TPP/Conversion/ConvertLinalgToXsmm/ConvertLinalgToXsmm.cpp
@@ -9,11 +9,11 @@
 #include "TPP/Dialect/Tpp/TppUtils.h"
 #include "TPP/Dialect/Xsmm/XsmmOps.h"
 #include "TPP/Dialect/Xsmm/XsmmUtils.h"
+#include "TPP/IR/MatcherUtils.h"
 #include "TPP/IR/StructuredOpMatcher.h"
-#include "TPP/MatcherUtils.h"
 #include "TPP/Passes.h"
 #include "TPP/TransformUtils.h"
-#include "TPP/Transforms.h"
+#include "TPP/Transforms/Transforms.h"
 #include "TPP/VNNIUtils.h"
 #include "TPP/ValueUtils.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
