+++ lib/TPP/IR/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_mlir_library(TPPIR
+  MatcherUtils.cpp
   StructuredOpMatcher.cpp
 
   ADDITIONAL_HEADER_DIRS
