+++ lib/TPP/Transforms/LinalgDeGeneralize.cpp
@@ -11,7 +11,7 @@
 #include "TPP/IR/StructuredOpMatcher.h"
 #include "TPP/Passes.h"
 #include "TPP/TransformUtils.h"
-#include "TPP/Transforms.h"
+#include "TPP/Transforms/Transforms.h"
 #include "mlir/Dialect/Arith/IR/Arith.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/Linalg/IR/Linalg.h"
