+++ lib/TPP/Transforms/LowerPacksAndUnpacks.cpp
@@ -8,7 +8,7 @@
 
 #include "TPP/Passes.h"
 #include "TPP/TransformUtils.h"
-#include "TPP/Transforms.h"
+#include "TPP/Transforms/Transforms.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/Linalg/IR/Linalg.h"
 #include "mlir/Dialect/Linalg/Transforms/Transforms.h"
