+++ lib/TPP/Transforms/TileConsumerAndFuseProducers.cpp
@@ -9,7 +9,7 @@
 #include "TPP/Dialect/Tpp/TppOps.h"
 #include "TPP/Passes.h"
 #include "TPP/TransformUtils.h"
-#include "TPP/Transforms.h"
+#include "TPP/Transforms/Transforms.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/Linalg/IR/Linalg.h"
 #include "mlir/Dialect/Linalg/Transforms/TilingInterfaceImpl.h"
