+++ test/TestLib/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_mlir_library(TPPTestLib
+  TestForToForAllRewrite.cpp
+  TestMatchers.cpp
+  )
